pyproject.toml
[project]
name = "ejemplos-lc-lg"
version = "0.1.0"
description = "ejemplos simples sobre LangChainy LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-google-community[gmail]>=2.0.7",
    "langchain-google-genai>=2.1.8",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "langgraph-supervisor>=0.0.27",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]
